requirements.txt
tabulate
pandas
scikit-learn
numpy
matplotlib
seaborn
xgboost
seaborn
colorama
plotly
fastparquet==2024.5.0
pyarrow==17.0.0
papermill==2.6.0
shap==0.46.0
python-docx